It's the correctly working turn client to test our restund server. The server must be run without auth.so module. This program should stay on a separate branch, cause I had to modify Thread class and it's not a bugfix but rather a dirty hack.
The real problem is the race condition that exists when the last
thread but one calls MessageQueue::Quit(). Than the last thread
can't read anything from the socket.
This is known bug in libjingle and is described here:
http://code.google.com/p/libjingle/issues/detail?id=83

Next step is to parametrize some variables, so we can launch several
of this test client from the shell.