Suppressions for memcheck errors in libjingle_peerconnection_unittest
The tests that fails under memcheck are disabled.
The suppressions reported are suppressed so we can start running the
test on the bot to catch at least future breakages.

Note that for the gtest_exclude file to be picked up, the
launch script must be run with the -b out/Release flag.

BUG=1204
TEST=tools/valgrind-libjingle/libjingle_tests.sh -b out/Release -t
libjingle_peerconnection_unittest

R=wu@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/1717005

git-svn-id: http://libjingle.googlecode.com/svn/trunk@353 dd674b97-3498-5ee5-1854-bdd07cd0ff33